Ported to new libfwk (Minor changes in math lib & naming)
- Notable change: disabled automatic conversion int* to/from float*

+++ src/editor/group_editor.cpp
@@ -216,11 +216,11 @@ namespace {
 
 			for(int n = 0; n < TileGroup::Group::side_count; n++) {
 				out.setViewPos(-center - worldToScreen(TileGroup::Group::s_side_offsets[n] * 9));
-				font->draw(out, int2(0, 0), Color::white, format("%d", m_tile_group->groupSurface(m_selected_group_id, n)));
+				font->draw(out, float2(0, 0), Color::white, format("%d", m_tile_group->groupSurface(m_selected_group_id, n)));
 			}
 				
 			out.setViewPos(-center +edit_rect.size() / 2);
-			font->draw(out, int2(0, 0), {Color::white}, format("setting surface: %d", m_selected_surface_id));
+			font->draw(out, float2(0, 0), {Color::white}, format("setting surface: %d", m_selected_surface_id));
 
 			/*
 			const char *names[] = {
@@ -243,7 +243,7 @@ namespace {
 		}
 
 		if(m_current_entry)
-			m_font->draw(out, int2(5, height() - 20), {Color::white, Color::black},
+			m_font->draw(out, float2(5, height() - 20), {Color::white, Color::black},
 					format("%s", m_current_entry->tile->resourceName().c_str()));
 	}
